+++ src/api/member.js
@@ -5,14 +5,14 @@ export function reissueAccessToken() {
 }
 
 export function fetchDeveloperList({
-                                     page = 0,
-                                     size = 10,
-                                     status,
-                                     nameInitial,
-                                     sortBy,
-                                     sortDir,
-    keyword,
-                                   } = {}) {
+  page = 0,
+  size = 10,
+  status,
+  nameInitial,
+  sortBy,
+  sortDir,
+  keyword,
+} = {}) {
   const query = { page, size };
 
   if (status) query.status = status;
@@ -24,31 +24,32 @@ export function fetchDeveloperList({
   return api.get("/members", { params: query });
 }
 
-export function searchDeveloperList({
-                                        keyword,
-                                        page = 0,
-                                        size = 10
-                                    }) {
-    console.log("search 호출, "+keyword);
-    return api.get("/members/search", {
-        params: {
-            keyword,
-            page,
-            size
-        }
-    });
+export function searchDeveloperList({ keyword, page = 0, size = 10 }) {
+  console.log("search 호출, " + keyword);
+  return api.get("/members/search", {
+    params: {
+      keyword,
+      page,
+      size,
+    },
+  });
 }
 
 export const updateMemberStatus = (employeeId, status) => {
-    return api.patch(`members/${employeeId}/status`, {
-        status: status,
-    });
+  return api.patch(`members/${employeeId}/status`, {
+    status: status,
+  });
 };
 
 export function fetchTechStacksByEmployeeId(employeeId) {
-    return api.get(`/members/${employeeId}/tech-stacks`);
+  return api.get(`/members/${employeeId}/tech-stacks`);
 }
 
 export function fetchDeveloperDetail(employeeId) {
-    return api.get(`/members/${employeeId}`);
+  return api.get(`/members/${employeeId}`);
+}
+
+/* 회원 가입 */
+export function signup(payload) {
+  return api.post("/members/signup", payload);
 }
